Documenting all of Hanoi's functions
For when I inevitably need to redo it all again.

+++ Assets/Prefabs/Puzzles/Hanoi/HanoiBall.cs
@@ -3,11 +3,18 @@ using UnityEngine;
 
 namespace Prefabs.Puzzles.Hanoi
 {
+    /// <summary>
+    /// Internal representation of a ball, storing its size and position.
+    /// </summary>
     public class HanoiBall
     {
         private static readonly Dictionary<GameObject, HanoiBall> RegisteredBalls =
             new Dictionary<GameObject, HanoiBall>();
-
+        
+        /// <summary>
+        /// Fetch the HanoiBall that corresponds to the <c>ballObject</c>.
+        /// </summary>
+        /// <param name="ballObject"><see cref="GameObject"/></param>
         public static HanoiBall GetHanoiBall(GameObject ballObject)
         {
             try
@@ -20,7 +27,11 @@ namespace Prefabs.Puzzles.Hanoi
                     $"Collider collided with object '{ballObject.name}' that is not a registered ball");
             }
         }
-
+        
+        /// <summary>
+        /// Registers HanoiBalls to the system, linking them to their <see cref="GameObject"/>.
+        /// </summary>
+        /// <param name="balls"><see cref="HanoiBall"/>s to register.</param>
         public static void AddHanoiBalls(params HanoiBall[] balls)
         {
             foreach (HanoiBall ball in balls)
@@ -28,10 +39,18 @@ namespace Prefabs.Puzzles.Hanoi
                     Debug.LogWarning($"Ball '{ball.Object.name}' already registered to Hanoi Towers system");
         }
 
+        /// <summary>
+        /// Object the HanoiBall represents.
+        /// </summary>
         public GameObject Object { get; }
         public readonly int Weight;
-
-        public HanoiBall(GameObject ball, int ballWeight)
+        
+        /// <summary>
+        /// HanoiBall constructor.
+        /// </summary>
+        /// <param name="ball"><see cref="GameObject"/> the HanoiBall represents.</param>
+        /// <param name="ballWeight"><see cref="ushort"/> representing how heavy the ball is. Balls can not be placed on lighter ones.</param>
+        public HanoiBall(GameObject ball, ushort ballWeight)
         {
             Object = ball;
             Weight = ballWeight;
